LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.font.parse.basic.worker-expected.txt


FAIL 2d assert_equals: ctx.font === '20px serif' (got 20PX   SERIF[string], expected 20px serif[string]) expected "20px serif" but got "20PX   SERIF"
